document-parser-backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "document-parser"
version = "0.1.0"
description = "文档解析存储系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "pandas>=2.1.0",
    "pdfminer.six>=20221105",
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0",
    "mail-parser>=3.15.0",
    "msg-parser>=0.2.6",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-json-logger>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
